flink/data-generator/pyproject.toml
[tool.poetry]
name = "data-generator"
version = "0.1.0"
description = ""
authors = ["Kai.He <kai.he.0811@qq.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Faker = "^12.3.0"
kafka-python = "^2.0.2"

[tool.poetry.dev-dependencies]
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
